Added base extended latin characters to Main fonts
Summary:
We'd like to support rendering extended latin characters in
order to support rendering text from European languages within
\text{}.  Adding the base characters to Main-Regular, Main-Italic,
and Main-Bold, is the first step in doing so.  Eventually we'll
need a way to compose accented characted from unicode text, but
that will follow in a later diff.

Test Plan:
- run `make ttf eot woff woff2` in the MathJaxFonts docker image in KaTeX
- copy the font files out of the docker, open them up on Font Book and verify that they have the correct glyphs at the correct unicode code points

Reviewers: emily